+++ adapter/httplib/httplib_test.go
@@ -26,11 +26,13 @@ import (
 	"time"
 )
 
-const getUrl = "http://httpbin.org/get"
-const ipUrl = "http://httpbin.org/ip"
+const (
+	getURL = "http://httpbin.org/get"
+	ipURL  = "http://httpbin.org/ip"
+)
 
 func TestResponse(t *testing.T) {
-	req := Get(getUrl)
+	req := Get(getURL)
 	resp, err := req.Response()
 	if err != nil {
 		t.Fatal(err)
@@ -63,12 +65,10 @@ func TestDoRequest(t *testing.T) {
 	if elapsedTime < delayedTime {
 		t.Errorf("Not enough retries. Took %dms. Delay was meant to take %dms", elapsedTime, delayedTime)
 	}
-
 }
 
 func TestGet(t *testing.T) {
-
-	req := Get(getUrl)
+	req := Get(getURL)
 	b, err := req.Bytes()
 	if err != nil {
 		t.Fatal(err)
@@ -210,7 +210,7 @@ func TestWithSetting(t *testing.T) {
 	setting.ReadWriteTimeout = 5 * time.Second
 	SetDefaultSetting(setting)
 
-	str, err := Get(getUrl).String()
+	str, err := Get(getURL).String()
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -223,8 +223,7 @@ func TestWithSetting(t *testing.T) {
 }
 
 func TestToJson(t *testing.T) {
-
-	req := Get(ipUrl)
+	req := Get(ipURL)
 	resp, err := req.Response()
 	if err != nil {
 		t.Fatal(err)
@@ -250,12 +249,11 @@ func TestToJson(t *testing.T) {
 			t.Fatal("response is not valid ip")
 		}
 	}
-
 }
 
 func TestToFile(t *testing.T) {
 	f := "beego_testfile"
-	req := Get(ipUrl)
+	req := Get(ipURL)
 	err := req.ToFile(f)
 	if err != nil {
 		t.Fatal(err)
@@ -269,7 +267,7 @@ func TestToFile(t *testing.T) {
 
 func TestToFileDir(t *testing.T) {
 	f := "./files/beego_testfile"
-	req := Get(ipUrl)
+	req := Get(ipURL)
 	err := req.ToFile(f)
 	if err != nil {
 		t.Fatal(err)