+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add: Convenient way to generate mock object [4620](https://github.com/beego/beego/issues/4620)
 - Infra: use dependabot to update dependencies. [4623](https://github.com/beego/beego/pull/4623)
 - Lint: use golangci-lint. [4619](https://github.com/beego/beego/pull/4619)
+- Fix lint and format code in adapter/httplib dir [4648](https://github.com/beego/beego/pull/4648)
 - Chore: format code. [4615](https://github.com/beego/beego/pull/4615)
 - Test on Go v1.15.x & v1.16.x. [4614](https://github.com/beego/beego/pull/4614)
 - Env: non-empty GOBIN & GOPATH. [4613](https://github.com/beego/beego/pull/4613)